requirements.txt
Flask
requests
Pillow
Flask-CORS
yt-dlp
gunicorn
